fix(web): Switched to client-side issues fetching due to CI issues with GH API limitations
